requirements.txt
# Core FastAPI and server
fastapi==0.119.0
uvicorn[standard]==0.37.0

# Database
sqlalchemy
psycopg2-binary
alembic
pgvector==0.3.6

# Background jobs
redis
rq

# Data validation
pydantic
pydantic-settings
pydantic[email]

# Utilities
python-dotenv
httpx
python-multipart
pyyaml
pandas

# Auth
passlib[argon2]
python-jose[cryptography]

# AI/ML - NOTE: Install requirements-cpu.txt FIRST for CPU-only PyTorch
sentence-transformers==3.3.1
google-generativeai==0.8.5
openai

# NLP/PII
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# Database connectors
pymongo
psycopg[async]
requests
psycopg-binary
slowapi
fakeredis==2.21.1
fakeredis
syrupy
pytest-timeout
aiosqlite
